WIP refactoring, make it compile
