+++ Marlin/advi3pp_.h
@@ -25,6 +25,11 @@
 #ifndef ADV_I3_PLUS_PLUS_PRIVATE_H
 #define ADV_I3_PLUS_PLUS_PRIVATE_H
 
+// This is only to ensure that Jetbrains CLion is parsing code properly inside the IDE
+#ifdef __CLION_IDE__
+#define HAS_BED_PROBE 1
+#endif
+
 #include "advi3pp_bitmasks.h"
 #include "advi3pp_enums.h"
 #include "advi3pp.h"
@@ -33,20 +38,16 @@
 
 namespace advi3pp {
 
-class Printer_;
-
 using andrivet::Callback;
 using BackgroundTask = Callback<void(*)()>;
 using WaitCalllback = Callback<void(*)()>;
 
 // --------------------------------------------------------------------
-// PagesManager
+// Pages
 // --------------------------------------------------------------------
 
-struct PagesManager
+struct Pages
 {
-    explicit PagesManager(Printer_& printer);
-
     void show_page(Page page, bool save_back = true);
     void show_wait_page(const __FlashStringHelper* message, bool save_back = true);
     void show_wait_back_page(const __FlashStringHelper* message, WaitCalllback back, bool save_back = true);
@@ -63,7 +64,6 @@ private:
     void handle_lcd_continue();
 
 private:
-    Printer_& printer_;
     Stack<Page, 8> back_pages_{};
     Page forward_page_ = Page::None;
     WaitCalllback back_;
@@ -71,19 +71,13 @@ private:
 };
 
 // --------------------------------------------------------------------
-// Settings
+// Handler - Handle inputs from the LCD Panel
 // --------------------------------------------------------------------
-
-struct Task;
-struct Printer_;
-struct LCD_;
 
 struct Handler
 {
-    explicit Handler(Printer_& printer);
-
     void handle(KeyValue value);
-    void show(bool save_forward, bool save_back);
+    void show(bool save_forward = false, bool save_back = true);
     virtual void store_eeprom_data(EepromWrite& eeprom);
     virtual void restore_eeprom_data(EepromRead& eeprom);
     virtual void reset_eeprom_data();
@@ -91,22 +85,14 @@ struct Handler
 
 protected:
     virtual bool dispatch(KeyValue value);
-    void invalid(KeyValue value);
-
-    PagesManager& pages() const { return printer_.pages(); }
-    Task& task() const { return printer_.task(); }
-    Printer_& printer() const { return printer_; }
-
-private:
     virtual void save();
     virtual void back();
     virtual void do_backup();
     virtual void do_rollback();
-    virtual Page do_show() = 0;
+    virtual Page do_show();
     virtual void do_save();
 
-private:
-    Printer_& printer_;
+    void invalid(KeyValue value);
 };
 
 // --------------------------------------------------------------------
@@ -115,8 +101,6 @@ private:
 
 struct LoadUnload: Handler
 {
-    explicit LoadUnload(Printer_& printer);
-
 private:
     bool dispatch(KeyValue key_value) override;
     Page do_show() override;
@@ -149,8 +133,6 @@ struct Preset
 
 struct Preheat: Handler
 {
-    explicit Preheat(PagesManager& mgr);
-
     void store_eeprom_data(EepromWrite& eeprom) override;
     void restore_eeprom_data(EepromRead& eeprom) override;
     void reset_eeprom_data() override;
@@ -159,14 +141,14 @@ struct Preheat: Handler
 private:
     bool dispatch(KeyValue key_value) override;
     Page do_show() override;
-    void send_preset();
+    void send_presets();
     void previous();
     void next();
     void cooldown();
 
 private:
     static const size_t NB_PRESETS = 5;
-    Preset presets_[NB_PRESETS];
+    Preset presets_[NB_PRESETS] = {};
     size_t index_ = 0;
 };
 
@@ -176,12 +158,6 @@ private:
 
 struct Move: Handler
 {
-    explicit Move(PagesManager& pages);
-
-private:
-    bool dispatch(KeyValue key_value) override;
-    Page do_show() override;
-
     void x_plus();
     void x_minus();
     void x_home();
@@ -195,6 +171,10 @@ private:
     void e_minus();
     void all_home();
     void disable_motors();
+
+private:
+    bool dispatch(KeyValue key_value) override;
+    Page do_show() override;
     void move(const char* command, millis_t delay);
 
 private:
@@ -207,12 +187,11 @@ private:
 
 struct SdCard: Handler
 {
-    explicit SdCard(PagesManager& mgr, Task& task);
-
     void show_first_page();
 
 private:
     bool dispatch(KeyValue value) override;
+    Page do_show() override;
     void show_current_page();
     void get_file_name(uint8_t index_in_page, String& name);
     void up();
@@ -230,10 +209,9 @@ private:
 
 struct Print: Handler
 {
-    explicit Print(PagesManager& pages);
-
 private:
     bool dispatch(KeyValue value) override;
+    Page do_show() override;
     void stop();
     void pause_resume();
     void advanced_pause();
@@ -251,8 +229,6 @@ private:
 
 struct SdPrint: Print
 {
-    explicit SdPrint(PagesManager& pages);
-
 private:
     void do_stop() override;
     void do_pause() override;
@@ -266,8 +242,6 @@ private:
 
 struct UsbPrint: Print
 {
-    explicit UsbPrint(PagesManager& pages);
-
 private:
     void do_stop() override;
     void do_pause() override;
@@ -281,7 +255,9 @@ private:
 
 struct FactoryReset: Handler
 {
-    explicit FactoryReset(PagesManager& pages);
+private:
+    Page do_show() override;
+    void do_save() override;
 };
 
 // --------------------------------------------------------------------
@@ -290,7 +266,20 @@ struct FactoryReset: Handler
 
 struct ManualLeveling: Handler
 {
-    explicit ManualLeveling(PagesManager& pages);
+private:
+    bool dispatch(KeyValue value) override;
+    Page do_show() override;
+    void back() override;
+    void point1();
+    void point2();
+    void point3();
+    void point4();
+    void point5();
+    void pointA();
+    void pointB();
+    void pointC();
+    void pointD();
+    void leveling_task();
 };
 
 // --------------------------------------------------------------------
@@ -299,7 +288,18 @@ struct ManualLeveling: Handler
 
 struct ExtruderTuning: Handler
 {
-    explicit ExtruderTuning(PagesManager& pages);
+private:
+    bool dispatch(KeyValue value) override;
+    Page do_show() override;
+    void back() override;
+    void start();
+    void heating_task();
+    void extruding_task();
+    void finished();
+    void settings();
+
+private:
+    double extruded_ = 0.0;
 };
 
 // --------------------------------------------------------------------
@@ -308,77 +308,165 @@ struct ExtruderTuning: Handler
 
 struct PidTuning: Handler
 {
-    explicit PidTuning(PagesManager& pages);
-};
+    void finished();
 
-// --------------------------------------------------------------------
-// Sensor
-// --------------------------------------------------------------------
+private:
+    bool dispatch(KeyValue value) override;
+    Page do_show() override;
+    void pid_tuning_step2();
+};
 
 #ifdef ADVi3PP_BLTOUCH
 
+// --------------------------------------------------------------------
+// Sensor Settings
+// --------------------------------------------------------------------
+
 struct SensorSettings: Handler
 {
-    explicit SensorSettings(PagesManager& pages);
-
     void send_z_height_to_lcd(double height);
     void save_lcd_z_height();
 
+private:
+    Page do_show() override;
+    void save_z_height(double height);
+    void do_save() override;
+};
+
+// --------------------------------------------------------------------
+// Sensor Tuning
+// --------------------------------------------------------------------
+
+struct SensorTuning: Handler
+{
+    void g29_leveling_finished(bool success);
+
+private:
+    bool dispatch(KeyValue key_value) override;
+    Page do_show() override;
+    void leveling();
+    void g29_leveling_failed();
     void self_test();
     void reset();
     void deploy();
     void stow();
+    //void z_height();
 
 private:
-    void save_z_height(double height);
+    bool sensor_interactive_leveling_ = false;
 };
 
-struct NoSensor: Handler
+// --------------------------------------------------------------------
+// Sensor Grid
+// --------------------------------------------------------------------
+
+struct SensorGrid: Handler
 {
-    explicit NoSensor(PagesManager& pages);
+private:
+    Page do_show() override;
+    void do_save() override;
 };
 
-#else
+// --------------------------------------------------------------------
+// Sensor Z Height
+// --------------------------------------------------------------------
+
+struct SensorZHeight: Handler
+{
+private:
+    Page do_show() override;
+    void do_rollback() override;
+    void save() override;
+    void home_task();
+    void center_task();
+};
+
+#else // No sensor
 
 struct SensorSettings: Handler
 {
-    explicit SensorSettings(PagesManager& pages) {}
-
-    void send_z_height_to_lcd(double height) {}
+    void send_z_height_to_lcd(double) {}
 	void save_lcd_z_height() {}
 
-    void leveling() {}
-    void self_test() {}
-    void reset() {}
-    void deploy() {}
-    void stow() {}
-
-    void start_z_height() {}
+private:
+    Page do_show() override;
 };
 
-struct NoSensor: Handler
+struct SensorTuning: Handler
 {
-    explicit NoSensor(PagesManager& pages);
+    void g29_leveling_finished(bool) {}
+
+private:
+    Page do_show() override;
+};
+
+struct SensorGrid: Handler
+{
+private:
+    Page do_show() override;
+};
+
+struct SensorZHeight: Handler
+{
+private:
+    Page do_show() override;
 };
 
 #endif
 
 // --------------------------------------------------------------------
+// No Sensor
+// --------------------------------------------------------------------
+
+struct NoSensor: Handler
+{
+private:
+    Page do_show() override;
+};
+
+// --------------------------------------------------------------------
+// Features Settings
+// --------------------------------------------------------------------
+
+struct FeaturesSettings: Handler
+{
+protected:
+    void send_features();
+    void do_save() override;
+
+    Feature features_ = Feature::None;
+};
+
+// --------------------------------------------------------------------
 // Firmware Setting
 // --------------------------------------------------------------------
 
-struct FirmwareSettings: Handler
+struct FirmwareSettings: FeaturesSettings
 {
-    explicit FirmwareSettings(PagesManager& pages);
+private:
+    bool dispatch(KeyValue key_value) override;
+    Page do_show() override;
+    void do_save() override;
+    void thermal_protection();
+    void send_usb_baudrate();
+    void baudrate_minus();
+    void baudrate_plus();
+
+    uint32_t usb_baudrate_ = 0;
 };
 
 // --------------------------------------------------------------------
 // LCD Setting
 // --------------------------------------------------------------------
 
-struct LcdSettings: Handler
+struct LcdSettings: FeaturesSettings
 {
-    explicit LcdSettings(PagesManager& pages);
+private:
+    bool dispatch(KeyValue key_value) override;
+    Page do_show() override;
+    void dim();
+    void buzzer();
+    void buzz_on_press();
 };
 
 // --------------------------------------------------------------------
@@ -387,7 +475,9 @@ struct LcdSettings: Handler
 
 struct Statistics: Handler
 {
-    explicit Statistics(PagesManager& pages);
+private:
+    Page do_show() override;
+    void send_stats();
 };
 
 // --------------------------------------------------------------------
@@ -396,7 +486,17 @@ struct Statistics: Handler
 
 struct Versions: Handler
 {
-    explicit Versions(PagesManager& pages);
+    String get_lcd_firmware_version();
+    bool is_lcd_version_valid() const;
+    void get_advi3pp_lcd_version();
+    void send_versions();
+
+private:
+    bool dispatch(KeyValue key_value) override;
+    Page do_show() override;
+    void versions_mismatch_forward();
+
+    uint16_t lcd_version_ = 0x0000;
 };
 
 // --------------------------------------------------------------------
@@ -405,8 +505,6 @@ struct Versions: Handler
 
 struct PrintSettings: Handler
 {
-    explicit PrintSettings(PagesManager& pages);
-
     void baby_minus();
     void baby_plus();
 
@@ -428,14 +526,12 @@ private:
 struct Pid
 {
     float Kp_, Ki_, Kd_;
-    int temperature_;
+    uint16_t temperature_;
 };
 
 struct PidSettings: Handler
 {
 public:
-    explicit PidSettings(PagesManager& pages);
-
     void set(uint16_t temperature, bool bed);
     void next();
     void previous();
@@ -448,8 +544,8 @@ private:
 
 private:
     static const size_t NB_PIDs = 5;
-    Pid backup_;
-    Pid pid_[2][NB_PIDs];
+    Pid backup_ = {};
+    Pid pid_[2][NB_PIDs] = {};
     bool bed_ = true;
     size_t index_ = 0;
 };
@@ -460,15 +556,13 @@ private:
 
 struct StepSettings: Handler
 {
-    explicit StepSettings(PagesManager& pages);
-
 private:
+    Page do_show() override;
     void do_backup() override;
     void do_rollback() override;
-    Page do_show() override;
     void do_save() override;
 
-    float backup_[XYZE_N];
+    float backup_[XYZE_N] = {};
 };
 
 // --------------------------------------------------------------------
@@ -477,17 +571,15 @@ private:
 
 struct FeedrateSettings: Handler
 {
-    explicit FeedrateSettings(PagesManager& pages);
-
 private:
     void do_backup() override;
     void do_rollback() override;
     Page do_show() override;
     void do_save() override;
 
-    float backup_max_feedrate_mm_s_[XYZE_N];
-    float backup_min_feedrate_mm_s_;
-    float backup_min_travel_feedrate_mm_s_;
+    float backup_max_feedrate_mm_s_[XYZE_N] = {};
+    float backup_min_feedrate_mm_s_ = 0;
+    float backup_min_travel_feedrate_mm_s_ = 0;
 };
 
 // --------------------------------------------------------------------
@@ -496,18 +588,16 @@ private:
 
 struct AccelerationSettings: Handler
 {
-    explicit AccelerationSettings(PagesManager& pages);
-
 private:
     void do_backup() override;
     void do_rollback() override;
     Page do_show() override;
     void do_save() override;
 
-    uint32_t backup_max_acceleration_mm_per_s2_[XYZE_N];
-    float backup_acceleration_;
-    float backup_retract_acceleration_;
-    float backup_travel_acceleration_;
+    uint32_t backup_max_acceleration_mm_per_s2_[XYZE_N] = {};
+    float backup_acceleration_ = 0;
+    float backup_retract_acceleration_ = 0;
+    float backup_travel_acceleration_ = 0;
 };
 
 // --------------------------------------------------------------------
@@ -516,15 +606,13 @@ private:
 
 struct JerkSettings: Handler
 {
-    explicit JerkSettings(PagesManager& pages);
-
 private:
     void do_backup() override;
     void do_rollback() override;
     Page do_show() override;
     void do_save() override;
 
-    float backup_max_jerk_[XYZE];
+    float backup_max_jerk_[XYZE] = {};
 };
 
 // --------------------------------------------------------------------
@@ -533,34 +621,8 @@ private:
 
 struct Copyrights: Handler
 {
-    explicit Copyrights(PagesManager& pages);
-};
-
-// --------------------------------------------------------------------
-// Sensor Tuning
-// --------------------------------------------------------------------
-
-struct SensorTuning: Handler
-{
-    explicit SensorTuning(PagesManager& pages);
-};
-
-// --------------------------------------------------------------------
-// Sensor Grid
-// --------------------------------------------------------------------
-
-struct SensorGrid: Handler
-{
-    explicit SensorGrid(PagesManager& pages);
-};
-
-// --------------------------------------------------------------------
-// Sensor Z Height
-// --------------------------------------------------------------------
-
-struct SensorZHeight: Handler
-{
-    explicit SensorZHeight(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 // --------------------------------------------------------------------
@@ -569,7 +631,8 @@ struct SensorZHeight: Handler
 
 struct ChangeFilament: Handler
 {
-    explicit ChangeFilament(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 // --------------------------------------------------------------------
@@ -578,11 +641,14 @@ struct ChangeFilament: Handler
 
 struct EepromMismatch: Handler
 {
-    explicit EepromMismatch(PagesManager& pages);
-
     bool does_mismatch() const;
     void set_mismatch();
     void reset_mismatch();
+
+private:
+    Page do_show() override;
+
+    bool mismatch_ = false;
 };
 
 // --------------------------------------------------------------------
@@ -591,7 +657,8 @@ struct EepromMismatch: Handler
 
 struct Sponsors: Handler
 {
-    explicit Sponsors(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 // --------------------------------------------------------------------
@@ -600,7 +667,8 @@ struct Sponsors: Handler
 
 struct LinearAdvanceTuning: Handler
 {
-    explicit LinearAdvanceTuning(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 // --------------------------------------------------------------------
@@ -609,7 +677,8 @@ struct LinearAdvanceTuning: Handler
 
 struct LinearAdvanceSettings: Handler
 {
-    explicit LinearAdvanceSettings(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 // --------------------------------------------------------------------
@@ -618,7 +687,8 @@ struct LinearAdvanceSettings: Handler
 
 struct Diagnosis: Handler
 {
-    explicit Diagnosis(PagesManager& pages);
+private:
+    Page do_show() override;
 };
 
 
@@ -628,8 +698,6 @@ struct Diagnosis: Handler
 
 struct AdvancedPause: Handler
 {
-    explicit AdvancedPause(PagesManager& pages);
-
     void advanced_pause_show_message(AdvancedPauseMessage message);
 
 private:
@@ -639,7 +707,6 @@ private:
     void filament_inserted();
 
 private:
-    PagesManager& pages_;
     AdvancedPauseMessage last_advanced_pause_message_ = static_cast<AdvancedPauseMessage>(-1);
 };
 
@@ -691,8 +758,6 @@ private:
 
 struct Task
 {
-    explicit Task(Printer_& printer, PagesManager& pages);
-
     void set_background_task(BackgroundTask task, unsigned int delta = 500);
     void clear_background_task();
     void execute_background_task();
@@ -703,8 +768,6 @@ private:
     void set_next_update_time(unsigned int delta = 500);
 
 private:
-    Printer_& printer_;
-    PagesManager pages_;
     unsigned int op_time_delta_ = 500;
     millis_t next_op_time_ = 0;
     millis_t next_update_time_ = 0;
@@ -713,28 +776,35 @@ private:
 
 
 // --------------------------------------------------------------------
-// LCD implementation
+// Main class
 // --------------------------------------------------------------------
 
-//! Implementation of the Duplication i3 Plus LCD
-struct LCD_
+struct ADVi3pp_
 {
-    explicit LCD_(PagesManager& pages);
+    void setup();
+    void idle();
+    void store_eeprom_data(eeprom_write write, int& eeprom_index, uint16_t& working_crc);
+    void restore_eeprom_data(eeprom_read read, int& eeprom_index, uint16_t& working_crc);
+    void reset_eeprom_data();
+    uint16_t size_of_eeprom_data() const;
+    void eeprom_settings_mismatch();
+    void temperature_error(const __FlashStringHelper* message);
+    bool is_thermal_protection_enabled() const;
+    void process_command(const GCodeParser& parser);
+    void advanced_pause_show_message(AdvancedPauseMessage message);
+    void set_brightness(int16_t britghness);
+    uint16_t last_used_hotend_temperature() const { return last_used_hotend_temperature_; }
+    uint16_t last_used_bed_temperature() const { return last_used_bed_temperature_; }
+    void save_settings();
+    bool is_busy();
 
-    static LCD_& instance();
-
-    void update();
-    void init();
     bool has_status();
     void set_status(const char* message);
+    void set_status(const __FlashStringHelper* fmt, ...);
     void set_status_PGM(const char* message);
     void set_alert_status_PGM(const char* message);
     void status_printf_P(const char* fmt, va_list argp);
     void set_status(const __FlashStringHelper* fmt, va_list argp);
-    void buttons_update();
-    void reset_alert_level();
-    bool detected();
-    void refresh();
     const String& get_message() const;
     void queue_message(const String& message);
     void reset_message();
@@ -747,78 +817,24 @@ struct LCD_
     void enable_buzz_on_press(bool enable);
     void buzz(long duration, uint16_t frequency = 0);
     void buzz_on_press();
+    uint32_t get_current_baudrate() const { return usb_baudrate_; }
+    void change_usb_baudrate(uint32_t baudrate);
+    Feature get_current_features() const { return features_; }
+    void change_features(Feature features);
 
 private:
     void buzz_(long duration);
-
-private:
-    PagesManager& pages_;
-    String message_;
-    String progress_name_;
-    mutable String progress_percent_;
-    mutable int percent_ = -1;
-    bool buzzer_enabled_ = true;
-    bool buzz_on_press_enabled_ = false;
-};
-
-// --------------------------------------------------------------------
-// Printer implementation
-// --------------------------------------------------------------------
-
-//! Implementation of the Duplicator i3 Plus printer
-struct Printer_
-{
-    Printer_();
-
-    void setup();
-    void task();
-    void auto_pid_finished();
-    void g29_leveling_finished(bool success);
-    void store_eeprom_data(eeprom_write write, int& eeprom_index, uint16_t& working_crc);
-    void restore_eeprom_data(eeprom_read read, int& eeprom_index, uint16_t& working_crc);
-    void reset_eeprom_data();
-    uint16_t size_of_eeprom_data() const;
-    void eeprom_settings_mismatch();
-    void temperature_error(const __FlashStringHelper* message);
-    bool is_thermal_protection_enabled() const;
-    void process_command(const GCodeParser& parser);
-    void advanced_pause_show_message(AdvancedPauseMessage message);
-    void set_brightness(int16_t britghness);
-    uint16_t last_used_hotend_temperature() const;
-    uint16_t last_used_bed_temperature() const;
-
-    void save_settings();
-
-private:
     void init();
     void check_and_fix();
     void update_progress();
     void send_status_data(bool force_update = false);
     void send_gplv3_7b_notice(); // Forks: you have to keep this notice
     void send_sponsors();
-    void send_versions();
     void read_lcd_serial();
-    void send_stats();
     void show_boot_page();
-    void reset_messages_task();
-    bool is_busy();
-
-    void g29_leveling_failed();;
-
-    String get_lcd_firmware_version();
-    void get_advi3pp_lcd_version();
-    bool is_lcd_version_valid() const;
-
-    void send_features();
-    void send_usb_baudrate();
-    void change_usb_baudrate();
 
     void icode_0(const GCodeParser& parser);
 
-    friend LCD_& LCD_::instance();
-
-private:
-    // Actions
     void screen(KeyValue key_value);
     void show_temps();
     void show_print();
@@ -827,176 +843,20 @@ private:
 
     void print_command(KeyValue key_value);
 
-    void preheat(KeyValue key_value);
-    void cooldown();
-
-    void move(KeyValue key_value);
-    void show_move();
-    void move(const char* command, millis_t delay);
-    void move_x_plus();
-    void move_x_minus();
-    void move_x_home();
-    void move_y_plus();
-    void move_y_minus();
-    void move_y_home();
-    void move_z_plus();
-    void move_z_minus();
-    void move_z_home();
-    void move_e_plus();
-    void move_e_minus();
-    void move_all_home();
-    void disable_motors();
-    void move_back();
-
-    void statistics(KeyValue key_value);
-    void show_stats();
-    void stats_back();
-
-    void pid_tuning(KeyValue key_value);
-    void pid_tuning_step1();
-    void pid_tuning_step2();
-    void pid_tuning_cancel();
-
-    void manual_leveling(KeyValue key_value);
-    void leveling_home();
-    void leveling_point1();
-    void leveling_point2();
-    void leveling_point3();
-    void leveling_point4();
-    void leveling_point5();
-    void leveling_finish();
-    void manual_leveling_task();
-
-    void extruder_tuning(KeyValue key_value);
-    void show_extruder_tuning();
-    void start_extruder_tuning();
-    void extruder_calibration_settings();
-    void extruder_tuning_heating_task();
-    void extruder_tuning_extruding_task();
-    void extruder_tuning_finished();
-    void cancel_extruder_tuning();
-
-    void sensor_settings(KeyValue key_value);
-    void sensor_settings_show();
-    void sensor_settings_save();
-    void sensor_settings_cancel();
-
-    void sensor_tuning(KeyValue key_value);
-    void sensor_tuning_show();
-    void sensor_tuning_back();
-    void sensor_leveling();
-    void sensor_z_height();
-
-    void sensor_grid(KeyValue key_value);
-    void sensor_grid_show();
-    void sensor_grid_cancel();
-    void sensor_grid_save();
-
-    void sensor_z_height(KeyValue key_value);
-    void sensor_z_height_cancel();
-    void sensor_z_height_continue();
-    void z_height_tuning_home_task();
-    void z_height_tuning_center_task();
-
-    void no_sensor(KeyValue key_value);
-    void no_sensor_back();
-
-    void change_filament(KeyValue key_value);
-    void change_filament_show();
-    void change_filament_continue();
-
-    void firmware(KeyValue key_value);
-    void firmware_settings_show();
-    void firmware_settings_thermal_protection();
-    void firmware_settings_baudrate_minus();
-    void firmware_settings_baudrate_plus();
-    void firmware_settings_save();
-    void firmware_settings_back();
-
-    void lcd(KeyValue key_value);
-    void lcd_settings_show();
-    void lcd_settings_dimming();
-    void lcd_settings_buzzer();
-    void lcd_settings_buzz_on_press();
-    void lcd_settings_back();
-
-    void factory_reset(KeyValue key_value);
-    void show_factory_reset_warning();
-    void do_factory_reset();
-    void cancel_factory_reset();
-
-    void versions(KeyValue key_value);
-    void versions_show();
-    void versions_mismatch_forward();
-    void versions_back();
-
-    void sponsors(KeyValue key_value);
-    void sponsors_show();
-    void sponsors_back();
-
-    void copyrights(KeyValue key_value);
-    void copyrights_show();
-    void copyrights_back();
-
-    void eeprom_mimatch(KeyValue key_value);
-    void eeprom_mimatch_continue();
-
-    void linear_advance_tuning(KeyValue key_value);
-    void linear_advance_settings(KeyValue key_value);
-    void diagnosis(KeyValue key_value);
-
 private:
-    PagesManager pages_;
-    LCD_ lcd_;
-    Task task_;
-    uint16_t lcd_version_ = 0x0000;
-    Feature features_;
-    uint32_t usb_baudrate_;
-    uint32_t usb_old_baudrate_;
-    Dimming dimming_{};
-    Graphs graphs_;
-
-    LoadUnload load_unload_;
-    Preheat preheat_;
-    Move move_;
-    SdCard sd_card_;
-    FactoryReset factory_reset_;
-    ManualLeveling manual_leveling_;
-    ExtruderTuning extruder_tuning_;
-    PidTuning pid_tuning_;
-    SensorSettings sensor_settings_;
-    FirmwareSettings firmware_settings_;
-    NoSensor no_sensor_;
-    LcdSettings lcd_settings_;
-    Statistics statistics_;
-    Versions versions_;
-    PrintSettings print_settings_;
-    PidSettings pid_settings_;
-    StepSettings steps_settings_;
-    FeedrateSettings feedrates_settings_;
-    AccelerationSettings accelerations_settings_;
-    JerkSettings jerks_settings_;
-    Copyrights copyrights_;
-    SensorTuning sensor_tuning_;
-    SensorGrid sensor_grid_;
-    SensorZHeight sensor_z_height_;
-    ChangeFilament change_filament_;
-    EepromMismatch eeprom_mismatch_;
-    Sponsors sponsors_;
-    LinearAdvanceTuning linear_advance_tuning_;
-    LinearAdvanceSettings linear_advance_settings_;
-    Diagnosis diagnosis_;
-    SdPrint sd_print_;
-    UsbPrint usb_print_;
-    AdvancedPause pause_;
-
     bool init_ = true;
-    bool sensor_interactive_leveling_ = false;
-    double extruded_ = 0.0;
-    uint16_t last_used_temperature = 200;
+    uint32_t usb_baudrate_;
+    Feature features_;
+    uint16_t last_used_hotend_temperature_ = 200;
+    uint16_t last_used_bed_temperature_ = 200;
+    String message_;
+    String progress_name_;
+    mutable String progress_percent_;
+    mutable int percent_ = -1;
+    bool buzzer_enabled_ = true;
+    bool buzz_on_press_enabled_ = false;
 };
 
 }
 
 #endif //ADV_I3_PLUS_PLUS_PRIVATE_H
-
